requirement.txt
asgiref==3.5.2
Django==3.2
django-filter==22.1
django-htmx==1.12.2
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
